♻️ ScoringView hook으로 기능분리
