+++ extension/src/contentScript/hooks/index.ts
@@ -1,7 +1,8 @@
 import { useBadge } from './useBadge';
 import { useProfile } from './useProfile';
+import { useScoring } from './useScoring';
 import { useRandomRecommandProblem } from './useRandomRecommandProblem';
 import { useRanking } from './useRanking';
 import { useRecommandProblems } from './useRecommandProblems';
 
-export { useBadge, useProfile, useRandomRecommandProblem, useRanking, useRecommandProblems };
+export { useBadge, useProfile, useScoring, useRandomRecommandProblem, useRanking, useRecommandProblems };
